requirements.txt
acme==1.12.0
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
asyncio==3.4.3
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
ccxt==4.4.91
certbot==1.12.0
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.1.8
cloud-init==20.4.1
comm==0.2.2
ConfigArgParse==1.2.3
configobj==5.0.6
contourpy==1.3.0
cryptography==3.3.2
cycler==0.12.1
dateparser==1.2.1
dbus-python==1.2.16
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distro==1.5.0
distro-info==1.0+deb11u1
exceptiongroup==1.3.0
executing==2.2.0
fastjsonschema==2.21.1
flask==3.1.1
fonttools==4.58.4
fqdn==1.5.1
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.18.1
httpx==0.28.1
idna==2.10
importlib-metadata==8.7.0
importlib-resources==6.5.2
iotop==0.6
ipykernel==6.29.5
ipython==8.18.1
isc==2.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.1
josepy==1.2.0
json5==0.12.0
jsonpatch==1.25
jsonpointer==2.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter-client==8.6.3
jupyter-core==5.8.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter-server==2.16.0
jupyter-server-terminals==0.5.3
jupyterlab==4.4.4
jupyterlab-pygments==0.3.0
jupyterlab-server==2.27.3
kiwisolver==1.4.7
lxml==4.6.3
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mistune==3.1.3
more-itertools==4.2.0
multidict==6.5.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook-shim==0.2.4
numpy==2.0.2
nvidia-nccl-cu12==2.27.3
oauthlib==3.1.0
overrides==7.7.0
packaging==25.0
pandas==2.3.0
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
ply==3.11
prometheus-client==0.22.1
prompt-toolkit==3.0.51
propcache==0.3.2
psutil==7.0.0
ptyprocess==0.7.0
pure-eval==0.2.3
pycares==4.9.0
pycparser==2.22
pycryptodome==3.23.0
pycurl==7.43.0.6
pygments==2.19.2
PyICU==2.5
pyinotify==0.9.6
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.15.5
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-binance==1.0.29
python-dateutil==2.9.0.post0
python-debian==0.1.39
python-debianbts==3.1.0
python-dotenv==1.1.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML==5.3.1
pyzmq==27.0.0
referencing==0.36.2
regex==2024.11.6
reportbug==7.10.3+deb11u2
requests==2.32.4
requests-toolbelt==0.9.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.26.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
systemd-python==234
ta==0.11.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250516
typing-extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
unattended-upgrades==0.1
uri-template==1.3.0
urllib3==1.26.5
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
xgboost==2.1.4
yarl==1.20.1
zipp==3.23.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.2.0
